operations: Allow rebase onto remote revision
Prior to this change, it was only possible to rebase a branch onto
another branch in the same repository.

This change implements the use of `fetchStartRevision` in rebase. When
the user provided a remote branch from a remote repository, that
revision is first fetched into the local repository before calling
gitaly-git2go.

This has as result the rebase command of gitaly-git2go no longer accepts
an upstream branch name, but it expects the revision where to rebase
upon instead.