Set a default backchannel registry in Gitaly server builder
testserver.RunGitaly currently doesn't set a default backchannel
registry in case on isn't provided. This is causing test failures
as using a nil registry causes panics when a multiplexed connection
is used. This is not valid use of the types so let's set a default
registry if none is provided.