support muxed connection in DisabledElector
GitLab's tests use Praefect with the DisabledElector. Recently an
MR was merged to default enable all Gitaly and Praefect feature flags
when the tests are being run. The pipeline was failign as the
DisabledElector did not support the connection multiplexing and the
feature flags were enabled. Fix this by passing the multiplexed connections
in to the DisabledElector as well.