doc: Document process for custom git patches
Now that we have a single distribution of git across all of GitLab's
official distributions of Gitaly, we are in a nicer position to add
custom patches to git. In order to ensure high standards and assert that
we won't ever start to diverge from upstream, this commit documents the
high bar we want to set ourselves for adding custom patches.