featureflag: Remove `AtomicFetch` feature flag
Atomic fetches have been tested in staging since last week and
production since the beginning of this week without any issues at all.

Remove the feature flag to default-enable the feature. While this may
feel fast, it was only added as a quick bailout in case it behaves
different from what was expected. It didn't, so the removal should be
safe.