operations: Add option to setup custom test transaction manager
The testhelpers to set up the operations service do not currently have
any way to override parts of the stack that is set up. As we'll have to
replace the transaction manager in the following commit with a custom
mocked one, this commit introduces a new `setupConfig` structure with
associated `setupOptions` which can be passed to both the relevant setup
functions. With the newly introduced `withTxManagerConstructor()`
option, this allows tests to override the transaction manager without
having to replicate the complete setup.