gittest: Do not precreate hook env in `WriteEnvToCustomHook()`
With `WriteEnvToCustomHook()`, we install a custom hook which prints the
hook environment variables into a file. This can be used either to
verify that the hook is run with the expected set of envvars, or to
check whether the hook ran in the first place.

The latter usecase is a bit convoluted though: because of how the setup
is done, we always precreate the file into which the envvars get printed
into. As a result, we cannot derive whether a hook ran based on the
existence of the env file, but instead must verify that it is still
empty.

Refactor `WriteEnvToCustomHook()` by using a temporary directory,
which easily allows us to not precreate the env file.