Propagate feature flags through gitaly-hooks
Hooks currently do not receive nor propagate the feature flags. This
is due to hooks being called through Git, which breaks the context
chain and thus the propagation. This commits propagate the feature
flags through the hooks to Gitaly's HookService. The feature flags
are added to the hook payload in order to pass them to the hooks.
The provided feature flags are then propagated via the outgoing context
to the HookService.