git2go: Explicitly ignore `CloseWithError()` return value
The golang-ci linter configuration currently has a rule to ignore error
return values of `CloseWithError()`. This rule kind of make sense given
that the only location where we call it is on an `io.PipeWriter()`, and
the documentation of that function explicitly states that it "always
returns nil.".

Let's explicitly ignore the error at the callsite and document why we do
so. This allows us to get rid of the rule.