repository: Fix silent error if sending GetInfoAttributes response fails
If `GetInfoAttributes()` doesn't have to do anything because there is no
gitattributes file, then we simply bail out and send a successful
response. We do not verify that sending the response succeeds though,
which could lead to us silently swallowing the error.

Fix the issue by returning the error of `stream.Send()` and drop the
corresponding linter ignore rule.