objectpool: Check errors when calling `pool.Remove()`
Our tests do not chekc the error value of `pool.Remove()` when executed
as part of test cleanup. This may easily hide any unexpected side
effects if e.g. the pool didn't exist or was removed because of whatever
bug.

Fix the issue by checking the returned error and drop the corresponding
linter ignore rule.