sql: Report errors when failing to clean SQL database
After tests, we're running `glsql.Clean()` to remove any data which
we've written to the database during tests. We never verify that the
call actually succeeds though, which would thus hide any errors in case
this ever started to fail.

Fix the issue by logging an error in case the function fails and drop
the corresponding linter ignore rules.