Remove config.Config from replicator_test.go
Multiple distinct gitaly services are used instead of
a single gitaly service with multiple storages.
The names of the storages and virtual storages were
re-defined to catch commonly used hardcoded "default"
value that may be a cause of false-positive test run.
The change also includes renaming of the package alias
gitaly_config to gconfig as the later used more commonly.
Couple of helper functions were inlined as there was only
a single place of usage.

Instead of creating a single gitaly service with multiple
storages test creates three distinct gitaly services as a
fleet of the single virtual storage of the praefect.
It represents actual production case of use.
The global config.Config replaced with local configs for
each service and operates independently. The use case
remain the same without any changes.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699