Remove config.Config from tempdir package
Usage of the global config.Config variable replaced with the
locally created value.

The order of the calls changed in the TestCleanTempDir test
as it started to create additional log entry about git
binary used for the test, because configuration of that happen
after log level change. By switching the order the old log
level is used to write that additional message and we have a
new level for the cleanTempDir function call.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699