remoterepo: Expose `gitalypb.Repository` interface
The `localrepo.Repo` type already implements the `repository.GitRepo`
interface, but the `remoterepo.Repo` doesn't. This often forces the user
to pass down both the remoterepo and its `*gitalypb.Repository`, which
is annoying.

Have the `remoterepo.Repo` embed its `*gitalypb.Repository` to fix this
weakness. In contrast to the localrepo, it doesn't only embed the
interface but the complete protobuf repository: this is because for
remote repos, it's quite likely that the user needs to call additional
RPCs, so they're tied together anyway.