StartGitalyServer to have ability to shutdown server manually
In some cases we need to terminate running gitaly service before
the test ends. RunGitalyServer only returns an address of the
running gitaly instance and we can't terminate it when we want.
To deal with it the StartGitalyServer is added. It returns
GitalyServer that provides an address of the running gitaly
and it has Shutdown method that allows to terminate running
instance when needed. If method was not called the service
will be terminated once the test is completed, so it doesn't
create a leak.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699