Remove config.Config from praefect service info
We are not dependent on the global config.Config var that
is why we are free to switch to more realistic test setups:
with two gitaly instances independent one from another.
The new WithShutdown option is used to stop one of the gitaly
services as there is no more direct reference to gRPC server
and we can't call Shutdown on it directly.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699