testhelper: Expose `WriteBlob()` function
While the testhelper already provides an interface to write multiple
blobs with indeterministic contents (the current time), this is not a
feasible interface for other tests which require deterministic contents.
This commit thus pulls out the logic to write the blob into a separate
function `WriteBlob()` which allows the caller to specify the blob's
contents.