conflicts: Fix use of ambiguous refs in ResolveConflicts
When resolving conflicts, we resolve the target branch via the
unqualified target branch name. But when the repository has both a
branch and tag with that name, then the query is ambiguous and would
return the resolved OID of the tag, not of the branch.

Fix the issue by converting branches to their fully qualified reference
names first.