conflicts: Fix fetching from target repository
When fetching a currently unknown revision from the target repository in
`ResolveConflicts()`, we first check whether the object already exists
in the source repo by calling `git rev-parse --verify $OID`. But as long
as the OID is validly formatted, this query will always return success
because git-rev-parse(1) only checks whether the revision can be
resolved to a valid OID, and if given an OID then this is always true.

Fix the issue by peeling the given OID to a commit, which causes
git-rev-parse(1) to resolve the actual object to see whether it's a
commit.