Add new test setup functions and use them
In order to break dependency on the global config.Config
variable this change introduces a set of new test setup
functions that are not dependent on the global variable.
As an example the archive_test.go file was modified to
rely on those new functions.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699