Refactor tests for CloneFromPoolInternal and CloneFromPool RPCs
The old test setup code is refactored to use the new setup functions.
The gitaly-ruby dependent test TestCloneFromPoolHTTP now runs
as a sub-test of the TestWithRubySidecar.
The package repository_test is not needed as with the new
setup functions there are no more circular dependency between
the packages.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699