Refactor fork_test.go
The old test setup code is refactored to use the new setup functions.
The newSecureRepoClient function moved to where it is used.
It is also marked as a test-helper. NewSecureRepoClient removed
as package changed from repository_test to repository, because
circular dependency is broken with new test setup functions.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699