Refactor fetch_test.go
The old test setup code is refactored to use the new setup functions.
The TestFetchFullServerRequiresAuthentication test removed
as it was just a sanity check for the test-setup code.
Package changed from repository_test to repository, because
circular dependency is broken with new test setup functions.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699