Refactor replicate_test.go
The old test setup code is refactored to use the new setup functions.
Package changed from repository_test to repository as there
is no more circular dependency issue because of the new test
setup functions.
The testserver.WithDisablePraefect configuration is used
because ReplicateRepository should not go through the praefect
proxy and must be a direct call to the gitaly server.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699