blob: Drop LFS pointer bitmap experiment
When searching for LFS pointers via ListLFSPointers, we need to walk
through the graph in order to answer the question which blobs are
reachable through the given set of references. This tends to become slow
when the repository's object graph is growing. In order to speed this up
for some cases, Gitaly has a feature flag that enables turning on use of
the bitmap index for this: instead of walking the graph, packfile
bitmaps have reachability for certain commits cached such that the
lookup becomes a lot faster.

This is the theory. In practice, enabling the use of bitmaps in
production has twice showed severe performance degradations. The first
time around a potential cause was identified with how negative tags
caused a pathological edge case. But given that we have backported this
patch and it's still showing a performance degradation, there must be
more issues lingering than only this.

Let's consider the experiment to be failed for now and remove both the
feature flag as well as the code it guards.