gitaly-git2go: Split out reading of conflict entries
In order to fix another bug with trailing newlines in the post-image of
a resolved merge conflict, we'll need to make available blob contents of
the conflicting sides. In order to avoid lookup of blobs multiple times,
this commit refactors the code and extracts `readConflictEntries()`:
given an index conflict, it returns conflict entries for all three
sides.

No functional change is expected from this commit.