conflict: Convert parser to accept conflict entries
In order to properly resolve conflicts, we'll need information about the
pre-images of the conflict entries. Refactor the parser to accept
conflict entries instead of only their paths as a preparatory step.

Note that this commit also changes the order of parameters: this is done
to more closely match other interfaces, which all have the ancestor as
first parameter. Further note that the testcase refactoring gets rid of
the differing paths given that all testcases use the same path for all
of ours, theirs and ancestor anyway.

No functional change is expected from this commit.