conflicts: Fix segfault in case unresolved conflicts have no ancestor
When resolving conflicts of an add-add conflict, then we do not have an
ancestor for this conflict. We still access the ancestor unconditionally
though in case that user-provided conflict resolutions do not solve all
existing conflicts when computing the error message.

Fix the issue by taking "our" path instead of the ancestor path if it
doesn't exist. "our" must always exist because there cannot be a
conflict without either "our" or "their" in a three-way merge.

This commit does not add a testcase because a later commit in this
series will add one that has inadvertently triggered the segfault.