Introduce gittest.Exec to exec git commands
The function testhelper.MustRunCommand designed for
running commands in the shell with reporting back result
of the execution. It server for too many purposes and
behaves differently in case 'git' command needs to be run.
Then it uses globally defined variable config.Config to
get path to the git binary. Most of the tests now
doesn't rely on that global variable and use one created
locally for each test.
In order to run only git related commands with locally defined
configuration the new gittest.Exec function is added.
As a use case example gittest.CreateRemoteBranch was changed
to use gittest.Exec function and local configuration.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699