ci: Add manual job to trigger QA
It's currently more complicated than necessary to trigger a QA run with
changes which are currently in review. As a result, we tend to just not
bother and wait for QA to crash at a later stage, when changes had been
merged to the master branch already and when GitLab is about to bump the
Gitaly server version. While we certainly don't want all MRs to be
tested by QA, it would be reasonable to do so for MRs which e.g.
deprecate existing features or which implement significant changes like
for example the recent Go ports.

Improve the situation by adding a new manual "qa" job. If executed, then
it triggers a QA pipeline with the tip of the MR as Gitaly server
version.