Introduce gittest.ExecStream to exec git commands
The function testhelper.MustRunCommand designed for
running commands in the shell with reporting back result
of the execution. It server for too many purposes and
behaves differently in case 'git' command needs to be run.
It also uses a globally defined variable config.Config to
get path to the git binary. Most of the tests now
doesn't rely on that global variable and use one created
locally for each test.
In order to run only git related commands with locally defined
configuration the new gittest.ExecStream function is added.
As a use case example gittest.CommitBlobWithName was changed
to use gittest.Exec and gittest.ExecStream function and local
configuration.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699