Replace testhelper.NewServer in blob package
Gitaly service setup done with testhelper.NewServer uses
a lot of additional code that is not part of the production.
It only creates a gRPC server and you need to manually
register all required services on it. Including creation of
all the dependencies required for those service.
The new testserver.RunGitalyServer function does most of
those operations for you. It uses mostly production code to
setup the server. It also provides you with meaningful
values for the dependencies you can use for services registration.