ci: Fix nightly jobs failing because git patches do not apply
Our nightly job tests against git's next and master branches to verify
that there's not going to be any new incompatibilities with the upcoming
git version. But since we have started to apply custom patches to git,
this nightly job has started to fail because the patch does not apply on
top of these branches.

Override GIT_PATCHES for this job such that we test against unpatched
versions of git.