command: Add test to detect goroutine leaks
Add TestNewNonExistent to help detecting goroutine leaks when
spawning a non-existent command with nil stderr.

When we `New` a command with nil stderr, a default stderr would
be created to help logging possible cmd errors, and in the
meanwhile we use io.Pipe to implement log size and line size
limiting. However, if we encounter an error from the latter
`cmd.Start()`, the PipeWriter will be left unclosed and related
goroutine leaks.

Signed-off-by: blanet <bupt_xingxin@163.com>