Fix racy context access in a flaky test
The `backchannel.ServerFactory` in the flaky test is checking a feature
flag at the same time from the context as the Praefect metadata is being
injected into few lines below. Fix the race by checkign the feature flag
status and using the result inside the ServerFactory instead of accessing
the context.