command: Refactor stderr handling
The stderr handling for command currently is kind of complicated
and very specifically designed for logrus. In some cases such as
spawning a non-existent command without caring about its stderr,
we even encounter goroutine leaks.

This commit refactors stderr handling for command by replacing the
`stderrCloser` and `stderrDone` with a single `stderrBuffer`.
`stderrBuffer` has similar functionality than before and itself
implements io.Writer, which can be passed directly to `cmd.Stderr`.
We no longer need to take much care about whether or not the pipe
is closed and whether or not any goroutine leaks.

Furthermore, we are able to use other logging libs that may not
provide the `io.Writer` interface as logrus does, and we can attach
pid or more useful tracking information with stderr output on error.

As we no longer create separate goroutines for logging, tests
involving stderr messages would get stuck(we are not reading from
the pipe when writing). In this commit we simply create an extra
goroutine as a temporary solution.

Signed-off-by: blanet <bupt_xingxin@163.com>