Fix data race on TestPostReceiveWithReferenceTransactionHook
After 77eab50b0 (Propagate feature flags through
gitaly-hooks, 2021-04-15) the problem with data
race was caught by the CI. The cause is that context
is used in a closure, but it is re-written in the
setup with InjectTransaction call. The fix is trivial
and requires closure creation to be moved after
InjectTransaction call.