Fix data race on ctx local variable re-write in sub-test
The ctx local variable is read in the closure for the
newMuxedOperationClient() setup. At the same time, there
is a re-write of that local variable during sub-test
execution. This causes a race condition.
As the closure only reads a feature flag set on the
context and doesn't depend on the values set during the
sub-test execution, let's fix the race condition by using
a new local variable inside the closure that is not
affected by the sub-test writes.