gitlab: Merge client and API implementations
The HTTPClient and the internal API implementation are currently split
up across two different times, even though the latter defers everything
to the HTTPClient. This separation of concerns is at times a bit weird,
and doesn't really help to understand the code.

Merge these two types into HTTPClient to lift this artificial
restriction.