remote: Fix detection of unknown remote HEAD
When FindRemoteRootRef isn't able to find a remote HEAD, then it's
supposed to fail with a NotFound error: that much is clear from the
implementation, where it's returning NotFound after the loop which is
iterating over lines printed by `git remote show`. But if Git isn't able
to determine the remote's HEAD, then the output won't be skipped, but it
instead prints `"(unknown)"`. We do not correctly determine this case
and just return this string as supposed remote HEAD to the caller.

Fix the issue by detecting this indicator and returning NotFound.