remote: Properly validate error conditions in testcases
The testcases verifying that `FindRemoteRootRef` properly validates
parameters do not verify that the test actually failed because of what
we thought they did. And of course, three of four of these tests don't:
they all fail because there is no remote specified and completely ignore
that there's no repository.

Fix the issue by tightening the tests such that they also verify error
messages. This uncovers that the RPC doesn't verify whether or not a
repository was passed, which is also fixed as part of this commit.