git: Move list of internal ref namespaces to central location
We have a bunch of internal reference namespaces which we need to treat
specially in some contexts. This list is currently defined in the
cleanup service as it is the only current user. Given that we're about
to introduce a second user, move this list into the `internal/git`
package to make it available for reuse in a central location.