prometheus: Extend HistogramVec interface
Many of our components use both the `Collect()` and `Describe()`
functions of Prometheus metrics in order to encapsulate metrics part of
a different struct and expose them via a simple combined interface. Our
MockHistogramVec type cannot be used in such contexts though because it
does not implement these functions.

Extend both our `HistogramVec` interface and the `MockHistogramVec`
implementation to fix this. While at it, document the exposed types and
functions.