gitlab: Split out and rename mock client
Mock client and the interface it is mocking are currently living in the
same file. This makes it hard to see what's what. Improve this by
splitting out the mock client into its own file. While we're at it, this
also renames the `mockAPI` into `MockClient`, which is more in line with
the `HTTPClient` implementation.