Makefile: Fix building git if PROFILE envvar is set
It was reported that building git fails with a bunch of compile errors
related to profiling. As it turns out, git's Makefile checks whether the
PROFILE environment variable is set and, if so, tries to build git with
profiling information. This is quite unexpected and most likely not what
the user building Gitaly wants, and with such a generic variable name
it's easy enough to accidentally hit.

Unset the environment variable to fix this issue. If a user really wants
to enable it, it should be added to the GIT_BUILD_OPTIONS array.