conflicts: Drop ResolveConflicts feature flag
While the ResolveConflicts RPC had been enabled in production for many
weeks now, we weren't yet able to default-enable it because that caused
various testcases in GitLab's QA tests to fail. These issues have now
been fixed, and QA started testing with all features default-enabled. So
we can be sure there are no more incompatibilities surfaced by QA.

Remove the feature flag. Given that we had it enabled in production for
such a long time, it doesn't feel like it would be required to first
default-enable it and then remove it, which would only cause us to add
one more release cycle on top of this topic.