remote: Add RemoteUrl parameter to FindRemoteRootRef
We're in the process of getting rid of on-disk remotes, instead of
asking the caller to always pass enough context, such that Gitaly is
able to construct an in-memory remote. FindRemoteRootRef is another of
these RPCs which currently only takes a remote name and thus relies on
on-disk state. This commit thus adds a new `RemoteUrl` parameter to the
RPC request such that callers can provide a URL instead of a name. To
enable authentication, we also add support for setting up HTTP headers.