gitaly-hooks: close gRPC ClientConn properly
`gitaly-hooks` does not close the connection it establishes to Gitaly
prior to exiting. Due to this, Gitaly does not receive a GOAWAY prior
to the connection terminating and thus can't know whether the connection
closed as planned or there was a connection error. This causes Gitaly then
to log errors when attempting to read from the broken connection. To fix
this, let's close the connection after it's been used for the HookService
calls. `gitaly-hooks` unfortunately exits the process from multiple
locations, including the `logger.Fatal`. Due to this, the closing is not
deferred but rather placed immediately after the HookService calls have
been performed.