Replace MustRunCommand with Exec in initRepoAt
Function initRepoAt used by InitBareRepoAt and
InitRepoWithWorktreeAtStorage to setup repositories
on the storage. By adding config.Cfg as an additional
parameter and replacing the MustRunCommand with Exec
we break dependency on the global config.Config variable.
The change also removes unused InitBareRepo and related
test for it.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699