Replace MustRunCommand with Exec in cloneRepo
Usages of the MustRunCommand replaced with Exec
in order to break dependency on the global
config.Config variable. All dependent functions
were changed as well. The same as the call site
for them.

In order to verify the list of the parameters used to
run git binary during service method execution the git
binary path replaced and that spyGitCfg used to run the
service. But the actual git still need to be used in other
cases. It was not an issue before as MustRunCommand uses
a global config.Config variable with proper git bin path set.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699