command: Fix improperly designed test for exported envs
`TestNewCommandProxyEnv` is designed to prove that http proxy related
envs can be whitelisted from environments and export to git spawn
later. But the current test just passed proxy related envs directly
into `New`, making the test meaningless.

Furthermore, a new test `TestNewCommandExportedEnv` is added to merge
`TestNewCommandTZEnv` and `TestNewCommandProxyEnv` together, as they
are actually testing the same thing.

Another test `TestNewCommandUnexportedEnv` is added as a negative
case proving that envs will not be exported to git spawn if they are
not on the white list.

Signed-off-by: blanet <bupt_xingxin@163.com>