Factor out stats collector cleanup
In https://gitlab.com/gitlab-org/gitaly/-/issues/3280 we learned that
it is important that the stats goroutine does not outlive the request
handler. The best way to ensure that is to use a defer statement, so
that we can be sure that we wait for the goroutine to finish on all
return paths from the handler.

This change moves the stats goroutine code around a bit so that we can
have a single defer statement, rather than a combination of
'pw.Close(); <-statsCh' we have to repeat over and over.