operations: Fix flaky UserRebaseConfirmable tests
Our tests exercising the UserRebaseConfirmable RPC are failing in CI for
the Go implementation. The failure is a simple race: we're awaiting that
the RPC wraps up and returns an EOF error, but it's not fast enough to
do so in one second. This most frequently happens in the code coverage
job, and most notably it only happens for the Go port. This is a bit
puzzling considering that the Go port is supposed to be faster given
that it's not creating a worktree like the Ruby implementation does. But
the fact that it is the code coverage job which is hitting the issue
hints that we hit in Go only because there's more instrumentation to be
able to determine which code paths have been hit, which naturally
doesn't exist in the Ruby port.

In any case, it doesn't really help to assert that the RPC finishes in a
fixed amount of time here. So let's fix flakiness by just dropping the
timeouts.