Replace NewTestGrpcServer with RunGitalyServer in gitaly services
It is a next iteration towards replacement of the old
test setup code with the new implementation that relies
more on the production code.
The t.Cleanup is used instead of defer to close client connections
with error verification.
Some functions were merged or inlined as had only a single usage.